src/weibo-emoji-update-checker/requirements.txt
onepush==1.1.1
python-dotenv==0.20.0